+++ changelog.md
@@ -1,1 +1,1 @@
-- fixed a rare server issue
+- fixed an issue with recipe ids all generating under same id